Update duckyinpython.py
This was a test to see if two payloads could be housed and launched.  Currently the payload that will run is based on what jumper is or is not connected.  Setup mode can still be activated by connecting GP0 to GND and GP5 to GND as well.  planning on using an 8 dip switch setup.  This should be able to house 7 payloads easily (if the PICO has enough memory.